+++ setup.py
@@ -1,27 +1,11 @@
-import codecs
-import os.path
-
 from setuptools import find_packages, setup
 
-
-def read(rel_path):
-    here = os.path.abspath(os.path.dirname(__file__))
-    with codecs.open(os.path.join(here, rel_path), "r") as fp:
-        return fp.read()
-
-
-def get_version(rel_path):
-    for line in read(rel_path).splitlines():
-        if line.startswith("__version__"):
-            delim = '"' if '"' in line else "'"
-            return line.split(delim)[1]
-    else:
-        raise RuntimeError("Unable to find version string.")
-
+with open("simuq/_version.py") as f:
+    version = f.readlines()[-1].split()[-1].strip("\"'")
 
 setup(
     name="SimuQ",
-    version=get_version("simuq/__init__.py"),
+    version=version,
     description="A Python package for quantum simulation with analog compilation",
     url="https://github.com/PicksPeng/SimuQ",
     license="BSD 3-Clause Clear",
@@ -51,5 +35,6 @@ setup(
             "qutip",
         ],
         "all": ["simuq[braket, ionq, dreal, qiskit, qutip]"],
+        "test": ["tox"],
     },
 )
