+++ simuq/qutip/__init__.py
@@ -1,0 +1,1 @@
+from simuq.qutip.qutip_provider import QuTiPProvider
