+++ simuq/transpiler.py
@@ -1,0 +1,3 @@
+class Transpiler:
+    def transpile(self, *args, **kwargs):
+        pass
